add support for mpv's rar demuxer
mpv has its own rar demuxer that works a bit differently than the
libarchive demuxer. One ugly thing about this change is that mpv's rar
demuxer uses "\" for paths instead of "/" (likely because most rar files
are created on windows). Some workarounds are added for this, but it is
likely to break rar files created on a *nix OS. Some of this stuff may
be reverted in the future if I can get mpv's rar demuxer to play nicer
with *nix paths.