make page stitching a little bit smarter
I'm still not 100% happy with it, but this is definitely an improvement
so I'll go ahead and push it. Stitching pages with multiple worker
scripts worked pretty dumbly before. It just had a constant offset
between each script and stitched pages together that way.

Instead, we create sort of a queue in manga-reader and then divide each
possible stitch combination evenly to every existing manga-worker
script. Doing this guarentees that multiple copies of the script never
attempt to stitch the same page combination. Of course, this makes the
offset paramater completely pointless now so it is removed. Users can
just up the pages option if they want to stitch a ton of pages all at
once.