move the init setup stuff out of start reader fn
There's no point in constantly redoing the initial scanning the
directory/generating filenames, etc. stuff everytime the the reader is
toggled on. Instead, we do this as soon as a file is loaded and
immediately unregister the event after it's done so the inital stuff is
only executed once. As a nice bonus, starting up the manga reader is
noticable faster now.