don't get stuck in create_stitches when quitting
Thanks to the manga*.lock files, it makes it possible to kill all worker
scripts by simply delete the lock files in the manga-reader script. Now
you no longer have to wait for the workers to finish stitching whatever
amount of pages are queued. Instead, the worker scripts finish the
current iteration they are on and then the loop is forcibly broken
meaning that you can quit much faster.