default to stitching pages all the way to the end
Thanks to f414cd5, the reader won't get stuck forever in the
create_stitches function waiting for all the pages to be stuck together.
So now we can finally default to just stitching as many pages together
as possible which is better for users and more efficient for the
scripts.