completely rewrite script based on lavfi-complex
This is a major change. Originally, I was dumb and didn't know about the
existence of the lavfi-complex fiter. So the script was written based
around imagemagick which is a pretty simple enough way to add pages to
each other. However, it is possible to just use the lavfi-complex option
in mpv to play multiple videos side by side. For the manga-reader, we
don't have to worry about the potential slowness of this since we're
only loading images.

lavfi-complex has several advantages over the old method. For one, it's
way, way faster. There's no need to write to the disk or anything ugly
like that. Adding another video track to mpv and displaying it with the
filter is nearly instant. Also, this means all the shell calls and path
ugliness is completely gone. So the script should now actually work on
Windows.

The only ugly thing in this implementation is a hacky sleep function
call that is basically used to ensure every file is loaded before
getting its dimensions. Otherwise, mpv switches through the playlist way
too fast and you get incorrect values.

Feature parity with the original is basically 99% there so I'm going to
merge it now. There's basically only a couple of small things missing
from the old one. First, double page mode doesn't have quite as
sophisticated width checks as it used to. If two images are the same
height, but their combined width would be way wider than the monitor's
aspect ratio, the images are combined and displayed here. On the old
one, it wouldn't do that. This rarely happens though, and it's an easy
enough fix (just lazy).

Continuous mode has one notable drawback now. Every image must be
exactly the same width. Double page mode has logic to handle mismatched
sizes so it's not an issue in that mode, but doing a similiar thing in
continuous mode while still supporting arbitrary page sizes is harder.
It's still possible, but it just means I have to add some more
boilerplate (which I don't feel like doing yet).

Overall, this way is a big improvement for me and I'm happy enough with
the state of this now. So let's go ahead and push it to master.