use pad filter for continuous mode
This should allow pages of different sizes to be stacked together in
continuous mode. In double page mode, we expect everything to be almost
exactly the same (which is why scale is used instead of pad). If two
pages are too different then we just choose only one. However in
continuous mode, these kinds of assumptions do not make sense because
the user just wants to scroll vertically through an archive. Therefore,
we can add black padding to the images as needed, center it, and then
vertically stack them.