fix subtle error in set_lavfi_complex_double
It turned out that the current logic in set_lavfi_complex_double was
actually flawed and in certain scenarios, the wrong page would be set to
scale resulting in the filter failing. Fix this by rewriting the
function so that it is always the external page that is scaled. Also
handle any errors that occur (this can still happen due to the imperfect
way page dimensions are saved) in a simplier, clearer manner.