fetch image dimensions on demand
Previously, the script had the logic of looping through the entire
directory to store all the required image dimensions to then perform its
logic. This was somewhat okayish but there would be a delay if you
happened to open up a large archive/directory which sucks. Instead,
rework the logic so that image dimensions are fetched on demand as you
progress through the directory/archive.