fix some indexing issues with validating pages
The logic in the last commit didn't take into account going back pages.
Make sure that the index in validate_pages starts from the index of the
page that change_page will switch to in the next call. Also check the
nil against valid_width instead of filedims since it is possible to get
filedims for a file but not its valid_width bool.