set some properties when the reader starts
Technically, this already was happening but instead I refactored it to a
couple of functions so the user's initial value before the reader
starts gets saved and restored properly. It was pointed out that
image-display-duration wasn't set to inf by default which is indeed a
little silly considering the purpose of this script so that was added in
here.