fix a remove_non_images regression
I mistakenly thought checking "" in string.match would return nil in
certain cases but it actually returns "". Instead if no extensions match
in the filename, check if the path has a "." at all in it. If not, we
assume it is a directory and keep it. If it does have the character,
then we assume it is a file with an invalid extension and remove it from
the playlist. Not perfect of course, but as long as you sanely name your
files and directories it's good enough.