fix some next/prev page issues in double page mode
It was a subtle issue. Basically the check whether to go back one or two
pages was too strict and would go back only one page if page dimensions
of the previous two pages were not fetched yet. Fix this by allowing the
double page conditional to work as long as the valid_width of the index
in question is not explicitly false (i.e. nil, the value when the page
dimensions have not be checked yet, is okay).