refactor next-page/prev-page/etc. logic
The logic between the different modes and going across pages got very
complex and very brittle as time went on as all sorts of weird edges got
piled on. This refactor should hopefully make it easier to understand
and easier to maintain. One major change is that much of the index
selection logic (for figuring out which page to change to next) was
ripped out of change_page and placed into the separate next_page,
prev_page, etc. functions. The reason is simply because what exactly
needs to be done varies greatly depending on whether you are going
backwards or forwards and cannot be easily generalized

One other change is that double_page always invokes ffmpeg's scale
filter now for simplification. The lavfi-complex filter needs the
heights of both images (in double page mode) to be exactly the same to
work. Now we just always scale one image to the other even if they are
already the same height. Presumably, ffmpeg scale just does nothing if
both images are the same height already, but I didn't bother to check
what exactly happens. Regardless, it still works the same.