rewrite double/continuous mode handling
Since lavfi-complex is used for double and continuous mode, the
dimensions of every page must be taken into account. The filter is only
valid in certain heights and widths. Previously, the mechanism for
figuring out the dimensions of every page was looping through the
playlist while waiting on mpv events. This mostly worked but still had
some race conditions and errors due to the nature of the wresting with
async.

Instead, this rewrite builds the mechanism using the on_preloaded lua
hook which offers several advantages over the old approach. Because the
hook executes before any tracks are selected, many calls to
lavfi-complex can be completely eliminated (no more flashing). Instead
of looping through files, the approach here is to add the desired pages
as additional video tracks, getting the width/height from the demuxer,
and storing that instead. Strictly speaking, the width/height from the
demuxer can be wrong according to the documentation but in nearly every
case it should be correct for our purposes here. If some edge case comes
up, we can work around it later.

Of course, this is a very big logic change so there's possibly some
regressions hidden in here, but overall it should be a massive
improvement. The page dimension measurements should always be accurate
now unless the demuxer is wrong (shouldn't happen often). A ton of logic
is simplified since the way the video tracks are added now guarantees
that all additional tracks are always external (unlike the old way which
was somewhat random with track selection). Fixes #10.