only add video tracks when neccessary
Not terribly important but a small optimization on the process here.
The prev_page and last_page functions have a special case in double page
due to how the logic of create_modes works. They must always check the
previous pages before actually changing playlist-pos to ensure that the
heights and widths are valid for displaying double pages. This check
only needs to be done once to store the file dimensions and after that
we can freely switch pages around. All this commit does is add look if
valid_width[new_index] is nil and if so, proceed to add the video tracks
and store the file dimensions. If we already have it, no need to waste
time adding the video tracks.