add a configurable zoom_multiplier option
By default, the zoom level in continuous mode tries to aim for exactly
one page (assuming that every stitched page is the same height). This of
course may not be the case or maybe someone would like a little
more/less zoom. Just add an additional factor and make it a configurable
option.