set a custom title for double/continuous modes
A cosmetic change. Both of these modes have multiple pages open so show
the title as "firstpage-lastpage.extension". Of course, this isn't
perfect and will break if your filenames have multiple "." in them
(you shouldn't do this though). Also this assumes everything in the
archive has the same file extension (not necessarily true).