properly unset lavfi-complex in double page mode
A regression from the rewrite that was unnoticed. If double pages were
being displayed and then the user navigates back to a page where double
pages are not supposed to be displayed (i. e. it's too wide), the script
did not unset the lavfi-complex filter. So double pages were still
awkwardly displayed. The valid_widths/similar_heights are properly being
tracked here so instead of simply returning from the
set_lavfi_complex_double, unset the filter. Also be sure to undo the
force-media-title property. We only need to do this if double pages are
currently being displayed, and we want to go back to displaying only one.