remove some deprecated code paths
The old "event.error" check was from mpv 32 or something around that
time. No point in carrying that around anymore so just get rid of it.
The script also made use of the new(ish) display-width and
display-height properties since May of this year. With the new 0.34
release, the old fallback code on monitor_height/monitor_width can be
removed. Those options should not be used anymore and users should
update their mpv to 0.34.