use the new image property
As of mpv 0.34, there is now a property for checking if the track is an
image. This is far better than the crude check the script did on its own
so just replace it with that. This effectively means that the
manga-reader script now requires at least mpv 0.34 (well technically
just that one specific commit) to work so go ahead and update the README
as well.