manga-reader: don't crap out if demux-h/demux-w isn't available
There's a while loop when trying to fetch this properties because they
can randomly fail, but some file formats geninuely won't have any values
in here. This just causes the manga reader to go into an enternal loop
which is no fun. Workaround this by setting an arbitrary limit when
trying to fetch filedims and then making up crap to deal with it. This
also require reworking the logic so the lavfi_complex filter doesn't
always unconditionally scale (probably an improvement). Maybe this
should just do an ffprobe command or something, but that can be for
later. Ugly but whatever.