manga-reader: add a hack for prev_page edge case
Unfortunately there seems to be an inherent race condition here which is
probably not trivial to properly fix. When going backwards, the manga
reader tries to check previously known dimensions so it knows if it can
go back 2 or 1 page (in double page mode only of course). If the
dimensions aren't known, it has to fetch them. This is where things get
bad because the added video tracks seem to not have a reliable order and
you can fetch incorrect dimensions. This results in an error when lavfi
complex is finally applied. Workaround this with a disgusting hack that
basically checks if you're going backwards, checks if there's bad
dimensions, clears them, assume a mismatch, and then step forward one
page. Probably fragile but it avoids some error messages and maybe might
have better behavior.