add an additional check for double page mode
Much better than the bad hack I thought of last commit, so that goes
away. When in double page mode, there are a couple of instances when we
do not know if the index is actually correct: prev_page and last_page.
In those cases, the dims of the pages have not been checked yet so we
have to first guess the index to go to. After on_preloaded hook checks
the relevant page dimensions we can then adjust the index and ensure we
select the correct page. This fixes a pretty long standing issue in a
decent enough and simple way.