don't automatically convert color pages to gray with lavfi-complex
This was a long standing quirk with the lavfi-complex filter. If one
page's format was gray and the other was color (e.g. yuv420p) it was
possible for the color page to be converted to gray depending on certain
conditions. We can fix this by also conditionally inserting a format
conversion filter if the pixel formats do not match and one of the pages
is gray. There are technically multiple gray formats out there
(gray16le, gray12le, etc.) and I'm not sure if anyone would
realistically every encounter one of these in the wild but convert those
as well to be on the safe side. This fix requires relatively new mpv
functionality (i.e. that I just merged today). On older versions,
getting the property will just fail so it's a no-op. Fixes #21.