convert gray pages to argb instead of matching the other page format
Blindly assuming the format of the color page is a good format to
convert to is not so good and can cause some weird formats. From testing
mpv's format conversion code, it turns out that the best conversion for
any gray format when paired with a non-gray format is always argb. So
instead of having all this logic with regards to comparing formats, just
always convert the page to argb. Annoyingly, ffmpeg can spam some
warnings about swscale and such. Just silence anything more verbose than
errors for ffmpeg while using this script. Fixes #30.